requirements.txt
diff-match-patch==20121119
Django==1.10.1
django-adminactions==1.5
django-classy-tags==0.8.0
django-geoposition==0.3.0
django-grappelli==2.11.1
django-import-export==1.0.0
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.6
openpyxl==2.5.0
pdfkit==0.6.1
Pillow==5.0.0
#pkg-resources==0.0.0
pytz==2018.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
